% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CSVML2}
\alias{CSVML2}
\title{C-SVM L2 - Support Vector Regression with C cost and L2 regularization.}
\arguments{
\item{y}{Vector with dependent variables should be -1 or +1. Dimension equal Nx1.}

\item{X}{Numeric matrix with the explanatory variables. Dimension equal NxP}

\item{C}{Cost parameter. Should be C>0.}

\item{kernel}{Name of the kernel that will be used.}

\item{parms}{Parameters associated with chosen kenel.}
}
\value{
List Support Vectors, Kernel used and parameters.
If the results for the Support Vectors are NaN it means that
there is no Support Vector and the Quadratic Programming Problem
is unfeasible.
}
\description{
Optimize the Lagrange multiplier for the C-SVM L2:

Min (1/2)u^{t}Qu-1^{t}u
s.t.
u>=0

where d is the vector of dependent variable,
and Q=(K+I/C).*(t(d)*d). C is the Cost parameter.
}
\examples{

A<-matrix(c(1,2,5,6,
2,4,1,2),nrow=4,ncol=2)
d<-c(-1,-1,+1,-1)
svm1<- CSVML2(d, A, 1, "Gaussian", c(0.5))

}

\references{
  \itemize{
    \item
      Chang Chih-Chung, Lin Chih-Jen\cr
      \emph{LIBSVM: a library for Support Vector Machines}\cr
      \url{http://www.csie.ntu.edu.tw/~cjlin/libsvm}

   \item
      Chih-Wei Hsu, Chih-Jen Lin\cr
       \emph{BSVM}
       \url{http://www.csie.ntu.edu.tw/~cjlin/bsvm/}
     \item
     J. Platt\cr
     \emph{Probabilistic outputs for support vector machines and comparison to regularized likelihood methods} \cr
     Advances in Large Margin Classifiers, A. Smola, P. Bartlett, B. Schoelkopf and D. Schuurmans, Eds. Cambridge, MA: MIT Press, 2000.\cr
     \url{http://citeseer.ist.psu.edu/viewdoc/summary?doi=10.1.1.41.1639}

    \item
    H.-T. Lin, C.-J. Lin and R. C. Weng\cr
    \emph{A note on Platt's probabilistic outputs for support vector machines}\cr
    \url{http://www.csie.ntu.edu.tw/~htlin/paper/doc/plattprob.pdf}

    \item
     C.-W. Hsu and C.-J. Lin \cr
     \emph{A comparison on methods for multi-class support vector machines}\cr
     IEEE Transactions on Neural Networks, 13(2002) 415-425.\cr
     \url{http://www.csie.ntu.edu.tw/~cjlin/papers/multisvm.ps.gz}

     \item
     K. Crammer, Y. Singer\cr
     \emph{On the learnability and design of output codes for multiclass prolems}\cr
     Computational Learning Theory, 35-46, 2000.\cr
     \url{http://webee.technion.ac.il/people/koby/publications/ecoc-mlj02.pdf}

     \item
     J. Weston, C. Watkins\cr
     \emph{Multi-class support vector machines} \cr
     In M. Verleysen, Proceedings of ESANN99 Brussels, 1999\cr
     \url{http://citeseer.ist.psu.edu/8884.html}
  }
}
\author{
  Pedro Albuquerque \email{pedroa@unb.br} \cr
  \url{http://pedrounb.blogspot.com/}
}

\seealso{
See \code{\link{.CallOctave}}, \code{\link{o_source}}, \code{\link{o_help}}
}

